@@ -1,7 +1,9 @@
-import js from '@eslint/js'
-import globals from 'globals'
-import reactHooks from 'eslint-plugin-react-hooks'
-import reactRefresh from 'eslint-plugin-react-refresh'
+import js from '@eslint/js';
+import globals from 'globals';
+import reactHooks from 'eslint-plugin-react-hooks';
+import reactRefresh from 'eslint-plugin-react-refresh';
+import react from 'eslint-plugin-react';
+import prettier from 'eslint-plugin-prettier';
 
 export default [
   { ignores: ['dist'] },
@@ -19,6 +21,8 @@ export default [
     plugins: {
       'react-hooks': reactHooks,
       'react-refresh': reactRefresh,
+      react,
+      prettier,
     },
     rules: {
       ...js.configs.recommended.rules,
@@ -28,6 +32,8 @@ export default [
         'warn',
         { allowConstantExport: true },
       ],
+      'prettier/prettier': 'warn',
+      'react/jsx-no-undef': 'error', // Warn for missing imports
     },
   },
-]
+];
